Change type from uint32 to int32 for moleculefeaturizer
torch.from_numpy cannot convert uint32. Force the array to be int32. This decreases the maximum allowed value by a factor of 2, but should be still large enough to cover the demand.

+++ chemprop/v2/featurizers/featurizers.py
@@ -24,12 +24,12 @@ class MorganFeaturizerMixin:
 
 class BinaryFeaturizerMixin:
     def __call__(self, mol: Chem.Mol) -> np.ndarray:
-        return self.F.GetFingerprintAsNumPy(mol)
+        return self.F.GetFingerprintAsNumPy(mol).astype(np.int32)
 
 
 class CountFeaturizerMixin:
     def __call__(self, mol: Chem.Mol) -> np.ndarray:
-        return self.F.GetCountFingerprintAsNumPy(mol)
+        return self.F.GetCountFingerprintAsNumPy(mol).astype(np.int32)
 
 
 @MoleculeFeaturizerRegistry("morgan_binary")
